Restructure to handle rbac credentials
